Move Counter.native, Counter.web to the containers folder
